pyproject.toml
[tool.poetry]
name = "core-synapse-ui"
version = "0.1.0"
description = "HTMX user interface for the Core Synapse plaftorm"
authors = ["devscouse <rob-jwebster@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
jinja2 = "^3.1.5"
requests = "^2.32.3"
pydantic-settings = "^2.8.1"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
